autofp: reduce flow storage space requirement
Use int16_t instead of a regular int to safe 2 bytes per flow.

+++ src/flow.h
@@ -326,7 +326,7 @@ typedef struct Flow_
     SC_ATOMIC_DECLARE(FlowRefCount, use_cnt);
 
     /** flow queue id, used with autofp */
-    SC_ATOMIC_DECLARE(int, autofp_tmqh_flow_qid);
+    SC_ATOMIC_DECLARE(int16_t, autofp_tmqh_flow_qid);
 
     uint32_t probing_parser_toserver_alproto_masks;
     uint32_t probing_parser_toclient_alproto_masks;
